requirements.txt
catboost==1.2.8
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.0
geographiclib==2.0
geopy==2.4.1
graphviz==0.20.3
joblib==1.5.0
kiwisolver==1.4.8
matplotlib==3.10.3
narwhals==1.40.0
numpy==2.2.6
packaging==25.0
pandas==2.2.3
pillow==11.2.1
plotly==6.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.2
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
watchdog==6.0.0
